docker-images/docker-madminer-ml/requirements.txt
h5py==2.10.0
madminer==0.7.1
numpy==1.18.3
pyyaml==5.3.1
